+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix bug with applying a cast expression to a view column.
 - Fix occasional memory errors caused by a lack of available mmap handles.
 - Fixed memory leak in groupby operations.
+- `names` parameter in Frame constructor is now checked for correctness.
 
 
 ### [v0.6.0](https://github.com/h2oai/datatable/compare/v0.6.0...v0.5.0) — 2018-06-05